A working reduce_while!
Needs more tests, and there are probably further efficiencies to be had
but this does work.

I still can't quite get my head around how / why we can't implement map
in reduce, but I think that because reduce just says "I will visit the
node at the end of each path and pass them to reduce_fun", as in not
every node gets passed to reduce_fn as we don't visit every node.
Whereas map requires that we pass all unvisited nodes as is to the
result and only visited ones get mapped as it were.

I suppose another way to implement it would be to have reduce visit every
node in the data structure, but only call the reducing fun if we are at
a place the path points to. That's tricky to do though because figuring
out if we are at a place that the path points to is not super easy AND
it would be way less efficient...

Anyway we still need to fix the Keyword Helper.